Adds code for controller, model, & routes
Adds code for the Beverage object:

Adds test code for for customer action to find all beverages by
  occasion

Adds code to the model to validate that all parameters for Beverages
are present, rating is an integer, and sets length limits for Name,
Description, and Recipe.

Adds code to routes for basic crud actions

Note: Beverage is still open.  Once crud actions are tested, the
controller will be set to protected.