Adds code to beverages_controller for lists
Adds and refactors code in the beverages_controller and routes.rb to
repurpose the index method and add a customer one.

We can now get a list of just the beverages for the current user.
However, to do this I changed the Index method to this:

  def index
    # @beverages = Beverage.all
    @beverages = current_user.beverages.all
    render json: @beverages
  end

Since this method 'knows' about current_user I repurposed it to get a
list of beverages that just belongs to the user.

I changed the method index_user_beverages to index_all_beverages and
gave it the basic code of:

  def index_all_beverages
    @beverages = Beverage.all
    render json: @beverages
  end